created test for pack function
