+++ requirements.txt
@@ -1,0 +1,16 @@
+atomicwrites==1.4.0
+attrs==21.1.0
+chardet==4.0.0
+colorama==0.4.4
+iniconfig==1.1.1
+memory-profiler==0.58.0
+numpy==1.19.4
+packaging==20.9
+pluggy==0.13.1
+psutil==5.7.3
+py==1.10.0
+Pympler==0.9
+pyparsing==2.4.7
+pytest==6.2.4
+PyYAML==5.4.1
+toml==0.10.2
